FW: create sandstone_system.h and signals.cpp
